+++ spec/models/dummy/dummy_instance_subscription_spec.rb
@@ -1,0 +1,8 @@
+require 'spec_helper'
+require Rails.root.join('../../spec/concerns/models/instance_subscription_spec.rb').to_s
+
+describe Dummy::DummySubscriber, type: :model do
+
+  it_behaves_like :instance_subscription
+
+end